Example data

Example dataset containing the catch, catch per unit effort (CPUE) and trawl survey abundance data for hake from New Zealand (Chatham Rise stock) between 1975 and 2012.

year

fishing year (i.e. 1975 is from September 1974 to August 1975)

landings

estimated landings in tonnes

survey

research survey biomass in tonnes

survey_cv

estimated research survey index coefficient of variation

cpue

standardised CPUE in tonnes per tow from the Eastern Chatham Rise

cpue_cv

estimated CPUE coefficient of variation from the standardisation

Usage

haknz

Source

Horn, P.L., Stock assessment of hake (Merluccius australis) on the Chatham Rise (HAK 4) and off the west coast of South Island (HAK 7) for the 2012-13 fishing year. New Zealand Fisheries Assessment Report, 2013/31: 58 p.