Plot posterior histograms

Plots histograms of posterior samples from an MCMC chain contained within a bdm class object.

Usage

histplot(object, ...)

# S3 method for class 'bdm'
histplot(
  object,
  pars = c("r", "logK", "lp__"),
  inc_warmup = FALSE,
  nrow = NULL,
  ncol = NULL,
  ...
)

Arguments

object

a bdm class object.

...

optional arguments to geom_histogram

pars

character vector of model parameters to be plotted. Defaults to pars = c('r','logK','lp__').

inc_warmup

logical value indicating whether MCMC warmup should be included in the plot.

nrow

passed to facet_wrap

ncol

passed to facet_wrap

Value

Returns a ggplot object that can be displayed or assigned and manuipulated using further arguments from the ggplot2 package.